Changed datatype of salary to float

+++ Management.h
@@ -8,12 +8,12 @@ class Staff{
 private:
 string name, id, gender, department, position, clock_in, clock_out;
 int year_started, age;
-double salary;
+float salary;
 
 public:
 Staff();
 
-Staff(string Name, int Age, string ID, string Gender,string Department, string Position, int Year_Started,int Salary, string in, string out);
+Staff(string Name, int Age, string ID, string Gender,string Department, string Position, int Year_Started,float Salary, string in, string out);
 
 void setName(string name);
 void setID(string id);
@@ -23,7 +23,7 @@ void setPosition(string position);
 void setClockIn(string clock_in);
 void setClockOut(string clock_out);
 void setAge(int age);
-void setSalary(int salary);
+void setSalary(float salary);
 void setYear(int year_started);
 
 string getName();
@@ -34,7 +34,7 @@ string getPosition();
 string getClockIn();
 string getClockOut();
 int getAge();
-double getSalary();
+float getSalary();
 int getYear();
 
 void displayProfile();
